Fix handling of shmKey in XvMCGetDRInfo
If we store the result of shmget in a CARD32 (unsigned int), then
checking if it returned -1 for an error by using >= 0 doesn't work.

Also, once the request is flushed from the buffer (as XReply does),
there's no guarantee the values in it are still valid, so it's better
to rely on our local variable instead.

Fixes: https://gitlab.freedesktop.org/xorg/lib/libxvmc/issues/1

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>